+++ .devcontainer/devcontainer.json
@@ -1,5 +1,6 @@
 {
 	"image": "mcr.microsoft.com/devcontainers/go:dev-1.22",
+	"name": "goland-operator-dev",
 	"features": {
 		"ghcr.io/devcontainers/features/docker-outside-of-docker:1": {}
 	},
