--- internal/controller/datadoggenericcrd/api_utils.go
@@ -1,27 +1,0 @@
-package datadoggenericcrd
-
-import (
-	"errors"
-	"fmt"
-	"net/url"
-
-	datadogapi "github.com/DataDog/datadog-api-client-go/v2/api/datadog"
-)
-
-func translateClientError(err error, msg string) error {
-	if msg == "" {
-		msg = "an error occurred"
-	}
-
-	var apiErr datadogapi.GenericOpenAPIError
-	var errURL *url.Error
-	if errors.As(err, &apiErr) {
-		return fmt.Errorf(msg+": %w: %s", err, apiErr.Body())
-	}
-
-	if errors.As(err, &errURL) {
-		return fmt.Errorf(msg+" (url.Error): %s", errURL)
-	}
-
-	return fmt.Errorf(msg+": %w", err)
-}
